Switched Debug Rendering to Modern OpenGL. Closes #22
Made a debug drawer class that can currently handle lines and points, but should be able to easily expand to use other types.
Removed irrelevant unbind in GameModel.
Added very basic shader for the debug drawer.
Disabled the networking components as it was slowing launch time, because it was looking for a server that was not available.
This patch also fixed the offset that was previously required for all bullet code, as it was caused by the legacy OpenGL.